Add test for Callback rule
