+++ src/Rule/Callback/Callback.php
@@ -20,7 +20,16 @@ final class Callback implements ParametrizedRuleInterface
         public bool     $skipOnEmpty = false,
         public bool     $skipOnError = false,
         public ?Closure $when = null,
-    ) {
+    )
+    {
 
+    }
+
+    public function getOptions(): array
+    {
+        return [
+            'skipOnEmpty' => $this->skipOnEmpty,
+            'skipOnError' => $this->skipOnError,
+        ];
     }
 }
